pyproject.toml
[project]
name = "gotogithub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.8.0",
]
